-moved Sex and PersonSearh1 to shared library and refactored code -moved REST URLs to constants
