+++ rest/src/test/java/ca/mikegabelmann/demo2/controller/rest/SexConverter.java
@@ -1,6 +1,6 @@
 package ca.mikegabelmann.demo2.controller.rest;
 
-import ca.mikegabelmann.demo2.persistence.model.Sex;
+import ca.mikegabelmann.demo2.codes.Sex;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
