+++ service/src/main/java/ca/mikegabelmann/demo2/service/PersonService.java
@@ -1,7 +1,6 @@
 package ca.mikegabelmann.demo2.service;
 
 import ca.mikegabelmann.demo2.dto.PersonDto;
-import ca.mikegabelmann.demo2.persistence.model.SexCode;
 
 import java.time.LocalDate;
 import java.util.List;
