+++ service/src/main/java/ca/mikegabelmann/demo2/service/PersonServiceImpl.java
@@ -1,8 +1,6 @@
 package ca.mikegabelmann.demo2.service;
 
 import ca.mikegabelmann.demo2.persistence.model.Person;
-import ca.mikegabelmann.demo2.persistence.model.Sex;
-import ca.mikegabelmann.demo2.persistence.model.SexCode;
 import ca.mikegabelmann.demo2.persistence.repository.PersonRepository;
 import ca.mikegabelmann.demo2.dto.PersonDto;
 import org.apache.logging.log4j.LogManager;
